Move helper methods into separate file and comment
